requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.3.2  # Older, more stable version
pymongo==4.6.1  # Compatible with motor 3.3.2
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2